Refactor Relayer Client to only use the api
