+++ packages/base/src/api/api.test.ts
@@ -9,7 +9,11 @@ const token = 'token';
 
 describe('createApi', () => {
   test('passes correct arguments to axios', () => {
-    createApi(key, token, apiUrl);
+    createApi({
+      apiKeyId: key,
+      apiUrl,
+      authenticationToken: token,
+    });
     expect(axios.create).toBeCalledWith({
       baseURL: apiUrl,
       headers: {
