+++ packages/base/src/api/auth.test.ts
@@ -1,17 +1,17 @@
-import { authenticate } from './auth';
+import { getAuthenticationToken, UserCredentials } from './auth';
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const { mockAuthenticateUser } = require('./__mocks__/amazon-cognito-identity-js');
 
 jest.mock('amazon-cognito-identity-js');
 
-const authData = {
-  Username: 'user',
-  Password: 'pwd',
+const authData: UserCredentials = {
+  apiKey: 'user',
+  apiSecret: 'pwd',
 };
 
-describe('authenticate', () => {
+describe('getAuthenticationToken', () => {
   test('passes correct arguments to Cognito', async () => {
-    await authenticate(authData, { ClientId: 'CLIENT', UserPoolId: 'POOL' });
+    await getAuthenticationToken(authData, { ClientId: 'CLIENT', UserPoolId: 'POOL' });
     expect(mockAuthenticateUser).toBeCalledWith(
       expect.objectContaining({
         username: 'user',
