+++ packages/relay/package.json
@@ -6,6 +6,7 @@
   "types": "./lib/index.d.ts",
   "scripts": {
     "build": "rm -rf lib && tsc",
+    "check": "tsc --noEmit",
     "test": "yarn test:unit",
     "test:unit": "jest --verbose",
     "watch": "tsc -w",
