+++ packages/relay/src/relayer/client.ts
@@ -1,0 +1,100 @@
+import { BaseApiClient } from 'defender-base-client';
+import { JsonRpcResponse } from 'web3-core-helpers';
+import { validateRelayerRequestPayload } from './validations';
+import { RelaySignerApiUrl } from '../api';
+import {
+  RelayerGetResponse,
+  RelayerTransactionPayload,
+  RelayerTransaction,
+  SignTypedDataPayload,
+  SignedMessagePayload,
+  SignMessagePayload,
+  ListTransactionsRequest,
+  RelayerParams,
+} from './types';
+
+export class Relayer extends BaseApiClient {
+  private jsonRpcRequestNextId: number;
+
+  public constructor(params: RelayerParams) {
+    super(params);
+    this.jsonRpcRequestNextId = 1;
+  }
+
+  protected getPoolId(): string {
+    return process.env.DEFENDER_RELAY_SIGNER_POOL_ID || 'us-west-2_iLmIggsiy';
+  }
+
+  protected getPoolClientId(): string {
+    return process.env.DEFENDER_RELAY_SIGNER_POOL_CLIENT_ID || '1bpd19lcr33qvg5cr3oi79rdap';
+  }
+
+  protected getApiUrl(): string {
+    return RelaySignerApiUrl();
+  }
+
+  public async getRelayer(): Promise<RelayerGetResponse> {
+    return this.apiCall(async (api) => {
+      return (await api.get('/relayer')) as RelayerGetResponse;
+    });
+  }
+
+  public async sendTransaction(payload: RelayerTransactionPayload): Promise<RelayerTransaction> {
+    validateRelayerRequestPayload(payload);
+    return this.apiCall(async (api) => {
+      return (await api.post('/txs', payload)) as RelayerTransaction;
+    });
+  }
+
+  public async replaceTransactionById(id: string, payload: RelayerTransactionPayload): Promise<RelayerTransaction> {
+    validateRelayerRequestPayload(payload);
+    return this.apiCall(async (api) => {
+      return (await api.put(`/txs/${id}`, payload)) as RelayerTransaction;
+    });
+  }
+
+  public async replaceTransactionByNonce(
+    nonce: number,
+    payload: RelayerTransactionPayload,
+  ): Promise<RelayerTransaction> {
+    validateRelayerRequestPayload(payload);
+    return this.apiCall(async (api) => {
+      return (await api.put(`/txs/${nonce}`, payload)) as RelayerTransaction;
+    });
+  }
+
+  public async signTypedData(payload: SignTypedDataPayload): Promise<SignedMessagePayload> {
+    return this.apiCall(async (api) => {
+      return (await api.post('/sign-typed-data', payload)) as SignedMessagePayload;
+    });
+  }
+
+  public async sign(payload: SignMessagePayload): Promise<SignedMessagePayload> {
+    return this.apiCall(async (api) => {
+      return (await api.post('/sign', payload)) as SignedMessagePayload;
+    });
+  }
+
+  public async query(id: string): Promise<RelayerTransaction> {
+    return this.apiCall(async (api) => {
+      return (await api.get(`txs/${id}`)) as RelayerTransaction;
+    });
+  }
+
+  public async list(criteria?: ListTransactionsRequest): Promise<RelayerTransaction[]> {
+    return this.apiCall(async (api) => {
+      return (await api.get(`txs`, { params: criteria ?? {} })) as RelayerTransaction[];
+    });
+  }
+
+  public async call(method: string, params: string[]): Promise<JsonRpcResponse> {
+    return this.apiCall(async (api) => {
+      return (await api.post(`/relayer/jsonrpc`, {
+        method,
+        params,
+        jsonrpc: '2.0',
+        id: this.jsonRpcRequestNextId++,
+      })) as JsonRpcResponse;
+    });
+  }
+}
